+++ packages/dso-toolkit/src/components/cursor-tooltip/cursor-tooltip.args.ts
@@ -1,0 +1,18 @@
+import { ArgTypes } from "storybook/internal/types";
+
+import { CursorTooltip } from "./cursor-tooltip.models.js";
+
+export interface CursorTooltipArgs {
+}
+
+export const cursorTooltipArgs: CursorTooltipArgs = {
+};
+
+export const cursorTooltipArgTypes: ArgTypes<CursorTooltipArgs> = {
+};
+
+export function cursorTooltipArgsMapper(a: CursorTooltipArgs): CursorTooltip {
+  return {
+    ...a
+  };
+}